iio: adc: ad4630: Support PGIA controll through scale attribute
IIO ABI is not clear with respect to the interface for input hardware gain.
Even though there is a definition for out_voltageY_hardwaregain attributes,
there is no counterpart for input.

The conventional (and intended) use of IIO ABI is to do
(_raw + _offset) * _scale to get from raw data to data in SI units
(or multiples of SI units).
Due to that, input gain must be taken into account by _scale properties
for the conversion to SI units be correct.

Remove custom gain attribute and allow user space to see and set PGIA
gains through in_voltage_scale_available and in_voltage_scale attributes.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>