iio: adc: ad4630: Set PGIA gain on probe
This will not solve the iio-osc bug of not writing to the voltage scale.
Though, writing to voltage scale through sysfs works so I don't understand
why it doesn't with iio-osc. It must be something silly like missing
some channel mask, maybe it's trying to write to in_voltage0_scale
instead of in_voltage_scale, or maybe I messed up and
overflown/underflow the scale value somehow.
I'm getting out of ideas of what it could be so maybe better take a
break and come back to it in one or two days.