iio: adc: adrv9002: add warmboot support
Warmboot allows to bypass the Initial calibrations and instead write a
set of known-good calibration coefficients directly to memory. This
improves significantly boot and profile loading times.

The way this is done is by adding a new binary attribute called
'warmboot_coefficients' and another device attribute called
'warmboot_coefficients_file'. The typical workflow looks like:

 1. cat warmboot_coefficients > /lib/firmware/current_profile_coeffs.bin
 2. echo current_profile_coeffs.bin > warmboot_coefficients_file
 3. cat current_profile.json > profile_config

So, looking at the above commands, one first needs to save the
coefficients for the desired profile (note that profile needs to be
loaded with warmboot disabled first so the initial calibrations can run
once). Then, life goes on and we might change to another profile but
when moving back to the same profile (now with warmboot enabled) one can
run 2 and 3 commands to load the profile. When setting up the device, we
see if warmboot is enabled and then we make use of the firmware API to
request the coefficients to load on the device.

Things are done like this mainly to save memory as these coefficients
buffers can be up to 1MB big and the main device object is already quite
big. Like this, we allocate memory for the coefficients only when needed
and free it as soon as we are done with it.

While at it, make sure to directly include 'mutex.h'.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>
(cherry picked from commit ac7c986e08ecd5f7fcda14b29f30b846ded76ab6)