ci: travis: run_build: don't allow dual licensed files
As of now, we were actually enforcing ADI code to be dual licensed.
Well, as seen recently that's more complicated than it looks like and
in order to get such code upstream, we might even need a lawyer to
sign in the patches. Hence, and as 99% of the time, we really not want
Dual licensed code, flip the logic and fail the job if new files are
dual licensed. While doing it some changes were done:

* Renamed the name of the job accordingly;
* Now we just check new files being added. Since they are being added
to ADI tree (and thus owned/maintained by ADI) we enforce this to all
files having them an 'Analog Devices' string or not;
* Fix the echo message displaying the wrong job name;
+ Also check the SPDX tag for multiple licenses;
* Added an helper to output the error messages (while massaging those
 messages).

Signed-off-by: Nuno Sa <nuno.sa@analog.com>
(cherry picked from commit 6885ebd5b836fe2e5c7f359cd21c1ace1e545321)